Switch to firefox-wayland
I have experiencing a bug with Firefox, in which on the first launch, it goes
to a black window. This seems to be due to a conflict between Firefox and the
GNOME-Wayland session. This can be fixed by using the `firefox-wayland`
package. This essentially sets the environment variable:
- `MOZ_ENABLE_WAYLAND=1`
The above could probably be done in a better way, and I would prefer for it to
be a little more flexible (by setting the environment variable depending on
whether or not I am using Wayland at that moment). Perhaps I will come back to
this in the future.