Update Python shell
Updated my typical python shell so that packages that rely on Numpy can still
work. Previously, Numpy packages cannot work with the previous shell duo to
missing shared libraries (seems to be something to do with the virtual
environment that I used). This one now works instead, and doesn't have the
same issues.